Added NS430 button
The N430 button will paste data compatible with the NS430 navaids.dat file to be used for flight planning or waypointing. This is useful for all aircraft that use the NS430 like
FC3 Modules
UH-1H
F-5E
and other aircraft that do not have a proper nav system installed in their craft.